feat(gvm): add lo

+++ cmd/activate.go
@@ -13,9 +13,11 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
+	"github.com/samber/lo"
 	"github.com/spf13/cobra"
 
 	"github.com/jaronnie/gvm/internal/global"
+	"github.com/jaronnie/gvm/internal/vm"
 )
 
 // activateCmd represents the activate command
@@ -28,16 +30,29 @@ var activateCmd = &cobra.Command{
 }
 
 func activate(cmd *cobra.Command, args []string) error {
+	cmd.SilenceUsage = true
+
 	gov := args[0]
 	if !strings.HasPrefix(gov, "go") {
 		return errors.New("invalid go version, please use gox.x")
+	}
+
+	rd := vm.NewReadDirVM()
+
+	vs, err := rd.List()
+	if err != nil {
+		return err
+	}
+
+	if !lo.Contains(vs, gov) {
+		return errors.Errorf("please exec `gvm install %s` first", gov)
 	}
 
 	goRoot := filepath.Join(global.GVM_CONFIG_DIR, gov)
 
 	_ = os.Remove(global.GVM_GOROOT)
 
-	err := os.Symlink(goRoot, global.GVM_GOROOT)
+	err = os.Symlink(goRoot, global.GVM_GOROOT)
 	if err != nil {
 		return err
 	}
